+++ test/durable_server/remote_placement_test.exs
@@ -47,7 +47,7 @@ defmodule DurableServer.RemotePlacementTest do
   end
 
   setup do
-    supervisor_name = :"test_supervisor_#{:erlang.unique_integer([:positive])}"
+    supervisor_name = unique_atom(:remote_placement_supervisor)
     prefix = "remote_placement_test_#{:erlang.unique_integer([:positive])}/"
 
     {:ok, supervisor_name: supervisor_name, prefix: prefix}